requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-multipart==0.0.6
pydantic==2.5.3
pyswisseph==2.10.3.2
